+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 20.11.11
 - the todo list CalDAV sync can now handle UTC dates from for example macOS Reminders or
   [OpenTasks](https://opentasks.app/) (for [#1966](https://github.com/pbek/QOwnNotes/issues/1966))
+- the disk database integrity check error handling was improved
 
 ## 20.11.10
 - the note tagging was improved (for [#1960](https://github.com/pbek/QOwnNotes/pull/1960), thank you @Waqar144)